requirements.txt
discord.py>=2.0
aiosqlite>=0.17
python-dotenv>=0.20
